backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
faiss-cpu==1.7.4
tiktoken==0.5.1
redis==5.0.1
pydantic==2.4.2
numpy==1.24.4
pytest==7.4.2
httpx==0.25.0